Add stub PPAPI for launching NaCl's sel_ldr process.
In ArcticSea, we're going to load the NaCl plugin in the render process.  The
NaCl plugin needs to create a sel_ldr process to host the NEXE, but when the
plugin lives inside the render process, the plugin is sandboxed, so it can't
create a process.  This API instructs the browser to launch a sel_ldr process
and wire up a bunch of IMC channels between the plugin and the new process.

In reality, the sel_ldr process is just chrome.exe run with a command line flag
that causes it to call NaClMain instead of BrowserMain or RendererMain. 
Previously, NaCl accomplished this task by smuggling a function pointer (cast as
an int) from render_process_impl to the NPAPI NaCl plugin.  Rather than use the
same approach again, we're creating a legit way for the PPAPI plugin to ask the
browser to spawn the sel_ldr process.

In BalticSea, we'll likely run the NaCl plugin out-of-process (e.g., in a PPAPI
process).  In that case, the plugin can just host the NEXE internally, and we'll
be able to remove this interface.

Review URL: http://codereview.chromium.org/5897004

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@69568 4ff67af0-8c30-449e-8e8b-ad334ec8d88c