ppapi_egl target is currently unused as verified by using:   grep --include="*.gyp*" -nr -F "ppapi_egl" .
However this library causes NaCl to require third_party/mesa which relative to NaCl is large.

This CL moves the definition of this static library from ppapi to the lib/gl directory under ppapi so that PPAPI users are not forced to DEPS in mesa.

Review URL: http://codereview.chromium.org/6342012

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@72090 4ff67af0-8c30-449e-8e8b-ad334ec8d88c