Revert 86625 - This change implements a first pass in the effort to remove the dependency of PlatformDevice within Chrome.  The Skia library now provides multiple back-ends for the SkDevice class, so PlatformDevice's inheritance of SkDevice, and the assumption of instances of PlatformDevice limits the use of these new back-ends.
A new set of helper functions is provided for the PlatformDevice entry points.  Upon construction of a PlatformDevice, a pointer to the interface is cached in the parent SkDevice's SkMetaData.  The new helper functions forward calls to the interface cached in the metadata.

BUG=NONE
TEST=NONE
Review URL: http://codereview.chromium.org/7019013

TBR=twiz@chromium.org
Review URL: http://codereview.chromium.org/6987019

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@86629 4ff67af0-8c30-449e-8e8b-ad334ec8d88c