Move Broker, Buffer, CharSet, and CursorControl to use the new thunk/API system.
This changes the way that the function APIs work in the webkit directory, now
they're created per-instance rather than globally, and have their
Plugininstance as a class member (this cleans up some code). I had to move the
InstanceData member to be a linked_ptr due to copy constructor issues, but this
should be more efficient anyway.
Review URL: http://codereview.chromium.org/7077024

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@87094 4ff67af0-8c30-449e-8e8b-ad334ec8d88c