Convert more interfaces to the new thunk system. This goes up to and including the ones starting with "F".
Since this adds a lot more interfaces, I added the macro stuff we used for the
old system to generate the various template specializations. This involded a
lot of renaming since the As* needs to match the name (I was previously leaving
off the "PPB_" part). I did other misc cleanup to the infrastructure.
Review URL: http://codereview.chromium.org/7082036

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@87415 4ff67af0-8c30-449e-8e8b-ad334ec8d88c