Don't copy the Graphics2D when binding a new one. This brings the implementation in line with the interface documentation (the old behavior would clobber any content).
The copying behavior was introduced to avoid flickering when resizing. To avoid
this, the PaintManager now doesn't bind the new device until it's been painted
to. This gives flicker-free painting.

This fixes a bug in the paint manager example that made it not paint.

TEST=manual
BUG=86568
Review URL: http://codereview.chromium.org/7215030

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@90386 4ff67af0-8c30-449e-8e8b-ad334ec8d88c