Add interfaces for requesting and receiving input event resources.
This converts the input event from a C struct to a resource to give us more
ability to change over time. This patch includes a proxy and a C++ wrapper for
this resource.

You now have to register for classes of input events. No events are sent by
default. This also allows us to specify whether the events support bubbling
or not, which allows us to better-optimize IPC.

TEST=none
BUG=none
Review URL: http://codereview.chromium.org/7285010

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@91711 4ff67af0-8c30-449e-8e8b-ad334ec8d88c