Missing changes for IDL
To minor changes to comments for the IDL.
Updated pp_stdint.idl so that we could generate it instead of checking in pp_stdint.h.
In addition it now provides a location for defining built in types.  Eventually we
can migrate type properties into this file.
Review URL: http://codereview.chromium.org/7448005

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@93125 4ff67af0-8c30-449e-8e8b-ad334ec8d88c