Cleanup in the file chooser API.
This revs the file chooser interface to no longer pass a separate options structure. This structure was weird because it only had two members, and didn't have ownership of the string which is potentially dangerous. The new interface is a bit easier to call.

I changed the string to take a Var, and the C++ layer now takes an Instance*
rather than an Instance& which is consistent with the other interfaces. I
updated the example.

TEST=manual
BUG=

Review URL: http://codereview.chromium.org/7660017

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@97558 4ff67af0-8c30-449e-8e8b-ad334ec8d88c