Remove webkit::ppapi::Resource.
This makes all resource _impl's derive directly from ppapi::Resource so we
can share code better. This means removing PluginInstances and converting
them to PP_Instances.

This adds a new ResourceHelper static class to help in the conversion of
resources to PluginInstances for the _impl classes.

Overall the new code is in general slightly worse than the old because using
the ResourceHelper is more annoying than just calling instance() on the old
webkit::ppapi::Resource object. However, I'm hoping that, because this will
allow us to move more code into shared_impl and reduce duplicate logic, it
will eventually have a net positive effect.

This also adds a ScopedPPResource class that works just like a scoped_refptr. We need this in a few places. Most of the places that used the old ScopedResourceId class could be removed now since resources have PP_Resource IDs generated even when there's no plugin reference (this didn't use to be the case) so we can pass resources as input params to the plugin even when there's no plugin ref, as long as there's a scoped_refptr to the Resource.

Review URL: http://codereview.chromium.org/7669055

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@98047 4ff67af0-8c30-449e-8e8b-ad334ec8d88c