More multi-version support (part 2)
Add member to Node to return a set of unique releases for that node.
Updates DefineStruct to iterate across all unique releases for the
struct, generating first the most recent release as we do today
and other releases without comments, and the version number 
appended to the structure name.

NOTE: This is work in progress. This CL adds support for
Interfaces that change which should be enough for M15.  While this
same CL creates multiple structures when the structures change, 
references to structures will not have mangled names which would 
cause errors.

BUG= http://code.google.com/p/chromium/issues/detail?id=89969
TEST= python generator.py --cgen --test --diff --range=M13,M14
Review URL: http://codereview.chromium.org/7751001

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@98298 4ff67af0-8c30-449e-8e8b-ad334ec8d88c