Add a new globals object for PPAPI tracking information.
[ Reland of 106142: http://codereview.chromium.org/8316008 ]

This adds a specialization on the host and plugin side of the proxy. This
replaces the ad-hoc singleton tracking done by the resource and var trackers
with just being getters on this global object.

Most code can use the single PpapiGlobals class. I also allow code to get the
host and plugin specializations since some code needs access to some specific
features of each side.

In a later pass I'll move the other stuff out of TrackerBase and delete it.

TEST=none
BUG=none
Review URL: http://codereview.chromium.org/8344025

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@106537 4ff67af0-8c30-449e-8e8b-ad334ec8d88c