Rename InterfaceID to ApiID and move the file.
This was originally in the proxy and had a 1:1 correspondence with an
interface. Then we reused this for other stuff and then merged some interfaces
into larger APIs (ppapi/thunk/*_api.h) so the name was no longer accurate.
It was wrong to be in the proxy directory since directories at a "lower level"
than the proxy (ppapi/shared_impl and webkit/plugins/ppapi) depended on it.

This renames to ApiID (I avoided APIID since it looks like a define) which is
the proper description of the class, and moved it to shared_impl. This fixes
the deps since there are no longer any bad dependencies on the proxy directory.

TEST=it compiles
BUG=none

Review URL: http://codereview.chromium.org/8333004

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@106619 4ff67af0-8c30-449e-8e8b-ad334ec8d88c