Rev the Flash interface to add new functionality.
This adds support for IDL in the private directory. Some of the interfaces there don't actually work, they were a work in progress and this breaks everything, so I moved the broken files to a new subdirectory. Everything left in the api/private directory can be autogenerated now.

BUG=
TEST=

Review URL: http://codereview.chromium.org/8930023

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@114534 4ff67af0-8c30-449e-8e8b-ad334ec8d88c