Switch a number of code and header files from DOS to Unix line-endings.
This is mainly relevant for the IDL-generated PPAPI headers, whose line-endings reflect the convention for the platform on which they were most recently re-generated, leading to bogus whole-file diffs.

Note that this CL specifically does not update the Copyright years on the affected files, since the changes are whitespace-only.

BUG=109116
TEST=diff --ignore-space-at-eol should give no output for this patch.
TBR=dmichael

Review URL: http://codereview.chromium.org/9087006

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@116311 4ff67af0-8c30-449e-8e8b-ad334ec8d88c