Use the new callback tracker and delete the old one
I added some ClearAndRun and ClearAndAbort helper functions to clean up the
common pattern of doing a swap to run the callbacks. I was also able to clean
up some of the file callbacks stuff since the parameters to the new
TrackedCallback are simpler.

This fixes a scary bug in the Transport API where all completion callbacks
would be aborted for all objects when you close the connection (!)

The HostResourceTracker is no longer necessary since the only point of this
was to integrate with the callback tracker. I removed it.

BUG=74062

TEST=

Review URL: http://codereview.chromium.org/9015009

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@116395 4ff67af0-8c30-449e-8e8b-ad334ec8d88c