Fix incorrectly resolved conflict, and add error message for missing sources.
When running the generator with default arguments from a CWD other than
the script's directory, the default arguments will be incorrect.  This adds
a check which will print an error message and fail.  If any argument, 
including the default arguments are provided on the command-line, then this
check is bypassed.

BUG= 109177
Review URL: http://codereview.chromium.org/9168014

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@117547 4ff67af0-8c30-449e-8e8b-ad334ec8d88c