Remove special handling for strings in var serialization.
Previously we had to have a lot of weird string handling because strings needed a PP_Module to be constructed, and because they had different methods for being created in the host and the plugin processes. We've removed all of these limitations, so now we can just delete the whole mess.

Review URL: https://chromiumcodereview.appspot.com/9316123

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@120724 4ff67af0-8c30-449e-8e8b-ad334ec8d88c