Fix nested invocation of graphics 2D Flush.
This fixes the Graphics 2D Flush ABORT callback to be a task instead of being
executed in a nested context. Previously, this would cause crashes because you
would get a paint message when destroying the paint manager.

This new way will issue the callback later. In the case of the paint manager,
the scoped callback factory will just drop the callback on the floor.

I added a check in the paint manager code just in case we add future failure
cases, we don't want to continue painting.

BUG=http://crbug.com/108688


Review URL: http://codereview.chromium.org/9404019

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@122234 4ff67af0-8c30-449e-8e8b-ad334ec8d88c