New file chooser interface that uses the new PP_ArrayOutput feature. This also changes PP_ArrayOutput to be pass-by-value.
This keeps backwards compat for the old interface. It fixes some bugs in the callback system that I found when working on the patch and adds some new machinery for doing array output in the proxy. It also re-enables the file chooser feature which was recently broken.

BUG=118857

Review URL: https://chromiumcodereview.appspot.com/9728001

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@129022 4ff67af0-8c30-449e-8e8b-ad334ec8d88c