If shared memory cannot be mapped, the audio thread should not be started as it will call the audio callback with an invalid buffer.
This also shifts all the state checks into the start of StartThread for safety/clarity. Note that this changes the behavior very slightly in that shared memory will be mapped even if a callback is not provided. However, AFAICT this will not be the common case and IMO simplifies the logic here. If you disagree, I'm happy to add the additional check.

BUG=118346
TEST=Audio tests


Review URL: http://codereview.chromium.org/10107021

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@134150 4ff67af0-8c30-449e-8e8b-ad334ec8d88c