Fix code in breakpad_win.cc which handles crash data entries for ppapi plugins.
Previously the code was broken for ppapi plugins. In particular, the URL and
plugin path entries were never added to the crash data for ppapi (both of these
fields wouldn't exist in any prior crashes).

The code for setting up this data structure is pretty error prone and
this problem only manifested itself inside GoogleUpdate (omaha). I think the code
here should be rewritten but this should be done as part of a subsequent CL.

This also re-enables URLs in ppapi crash reports on windows.

BUG=135480
TEST=Made sure ppapi crashes get uploaded when the url is included.
NOTRY=true

Review URL: https://chromiumcodereview.appspot.com/10704179

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@146643 4ff67af0-8c30-449e-8e8b-ad334ec8d88c