Add an IPC for PNaCl to check if the session is incognito, before deciding to cache a translated nexe (current cache mechanism won't work in incognito). Use this to decide whether or not to cache translations for now.
For now, the linker always writes to a self-deleting temp
file, instead of to the cache directly.  We then copy the
data from the temp file to the cache as a separate step.

Took about 60ms to copy a 4.6MB nexe in debug mode with a
128kb buffer and 15ms with a 1mb buffer...

BUG= http://code.google.com/p/nativeclient/issues/detail?id=2683
TEST= try compiling an app that doesn't use the File system
in incognito mode -- we should add a test though...
we should add a test though

Review URL: https://chromiumcodereview.appspot.com/10826107

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@149749 4ff67af0-8c30-449e-8e8b-ad334ec8d88c