Add a query function for the scaling feature.
We need to be able to query whether scaling is supported before creating a
Graphics2D object so we know which code path to use. This just adds a function
to do that checking.

I also renamed Graphics2DDev to Graphics2D_Dev for consistency with the other
Dev wrapper objects.

TEST=none
BUG=none

Review URL: https://chromiumcodereview.appspot.com/10832428

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@152829 4ff67af0-8c30-449e-8e8b-ad334ec8d88c