Switch OnMoreData() to use AudioBus.
As titled, with this change we're now piping float data around the pipeline from
end to end. This change is in preparation for browser side channel remixing and
resampling.

As a consequence of this change the shared memory now represents the
contents of an AudioBus object, which is essentially audio data in a float
planar format.

BUG=114700
TEST=Should be no audible change. Ran all existing tests. Compiled ran
WebAudio/HTML5/WebRTC on all platforms and PPAPI on Linux.


Review URL: https://chromiumcodereview.appspot.com/10832285

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@154951 4ff67af0-8c30-449e-8e8b-ad334ec8d88c