PNaCl: Remove Scons code for automatically generating pnacl_shim.c
This makes the Scons build consistent with the Gyp build.
pnacl_shim.c is currently checked in, and the Gyp build uses the
checked-in copy.  The checked-in copy is supposed to be regenerated by
scripts in ppapi/generators, and there is a presubmit check that
reminds committers to do the regeneration.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=3018
TEST=PNaCl is covered by nacl_integration on *_naclmore bots

Review URL: https://chromiumcodereview.appspot.com/10933002

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@157498 4ff67af0-8c30-449e-8e8b-ad334ec8d88c