Convert url request info to new proxy API.
This splits out the helper functions in the in-process version to a new file: url_request_info_util. This will be moved to content/renderer when we move the url_loader.

The unit test was moved from webkit/plugins/ppapi to content/renderer/ppapi. I made it a browsertest which allowed the removal of a little boilerplate.

I had to add "internal" functions for loading stuff with just the data struct rather than a resource so the proxy doesn't have to create an in-process URLRequestInfo resource (which can't be created now when running out-of-process) just to create a request. This should be a little more efficient now anyway, and these duplicates will go away when such code moves to the new proxy design.

BUG=

Review URL: https://codereview.chromium.org/10913257

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@157588 4ff67af0-8c30-449e-8e8b-ad334ec8d88c