NaCl: Remove "#ifdef NACL_STANDALONE" code paths
We no longer support building the NaCl PPAPI plugin as a standalone
plugin that can be loaded via --register-pepper-plugins: the
standalone NaCl build no longer tests that.

We simplify the code under the assumption that NACL_STANDALONE is
never defined.  This removes the plugin's support for launching the
standalone sel_ldr.

Remove NaClPluginGetSandboxISA(), which was only used in a test case
that has been removed.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=2730
TEST=trybots

Review URL: https://codereview.chromium.org/11274025

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@164106 4ff67af0-8c30-449e-8e8b-ad334ec8d88c