Add support in NaCl browser tests for testing pnacl-translated nexes
This CL enables bitcode builds of the various PPAPI libs,
and enables support PNaCl builds of several nexe test targets
(using pre-translated nexes for now).

It also adds a dependency on the IRT shim for those nexes,
and causes the shim to get built in-place for the pnacl toolchain. (since the SDK-packaged toolchain is in a different location, the SDK build script is updated so it copies the shim from the right place).

BUG= http://code.google.com/p/nativeclient/issues/detail?id=2484


Review URL: https://chromiumcodereview.appspot.com/11348048

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@167335 4ff67af0-8c30-449e-8e8b-ad334ec8d88c