Simplify platform_canvas.h by recognizing that PlatformCanvas does not actually extend SkCanvas in any way, other than provide a host of constructors (and delayed constructors in the form of 'initialize' methods).
These late initializers are a problem, as SkCanvas is deprecating its setDevice() call,
moving to model where the backingstore/device for the canvas must be created before the
canvas is created. This is necessary to allow skia to continue to extend SkCanvas for
its backends (e.g. GPU, PDF, Picture, Pipe, etc.).

The practical change in this CL is to make PlatformCanvas just a typedef for SkCanvas,
and change the call-sites that want to call initialize() to instead create the canvas
using one of the provided Factory functions (e.g. CreatePlatformCanvas). The modifier
Platform is maintained, to document that this canvas may be backed by platform-specific
pixels (e.g. allocated by GDI or cairo).
Review URL: https://codereview.chromium.org/11138024

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@167669 4ff67af0-8c30-449e-8e8b-ad334ec8d88c