Refactor the URLResponseInfo to use new design
This puts all of the URLResponseInfo attributes in a struct so it can be sent over IPC in one message rather than requiring one sync IPC per attribute access.

This includes a new example of streaming to a file that I used to do some manual tests of this change.

I added and improved some documentation that I noticed when trying to write the example.

Review URL: https://codereview.chromium.org/10993031

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@168273 4ff67af0-8c30-449e-8e8b-ad334ec8d88c