Refactor the way singleton-style resources are exposed via PPB_Instance
Currently a lot of boilerplate needs to be added everytime we add a new singleton style resource and as we add more of them this will accumulate. This patch reduces the amount of code needed by exposing a GetSingletonResource function in PPB_Instance_API which takes an ID specifying what type of resource is needed. A new Enter type, EnterInstanceAPI, is provided which is templated on the singleton API you want access to.

BUG=

Review URL: https://codereview.chromium.org/11359063

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@170221 4ff67af0-8c30-449e-8e8b-ad334ec8d88c