CDM/PPAPI: Fix plugin-side ref-counting leak
This makes the reference counting such that we always pass a reference to the plugin (in-process and out-of-process).

This also reverts https://src.chromium.org/viewvc/chrome?view=rev&revision=173626, since this is a more correct/complete fix. Now, if we end up trying to send the same HostResource more than once to the plugin, we should fail a DCHECK with this upcoming CL:
https://codereview.chromium.org/11576024/

BUG=165717,166498,166341

Review URL: https://codereview.chromium.org/11593032

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@173855 4ff67af0-8c30-449e-8e8b-ad334ec8d88c