Remove InstanceMessageFilter
This was previously used for handling PPB_Flash_Print IPCs which has now been moved to the pepper singleton-style resource model. The use of this class has largely been replaced by singleton-style resources and I don't forsee new uses of it in the near future. It can be added back in the future if needed.

BUG=


Review URL: https://chromiumcodereview.appspot.com/11645039

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@174905 4ff67af0-8c30-449e-8e8b-ad334ec8d88c