Breakpad: Add a test for crashes in the browser process
Use "--crash-test" to generate a crash in the browser process and
check that it produces a Breakpad crash dump.

This is in preparation for compiling Breakpad into Linux Chromium by
default, so that I will be able to check that crash_dump_tester.py
finds crash dumps OK on Linux.

We extend crash_dump_tester.py to launch both 32-bit and 64-bit
versions of crash_service.exe on Windows, so that we can get crash
dumps from the 32-bit browser process even on 64-bit systems.  We
introduce the cleanups_funcs list to make this cleaner.

This is not the best place to put Breakpad tests, but Chrome doesn't
have any other adequate framework for testing Breakpad at the moment.

BUG=105778
TEST=nacl_integration

Review URL: https://codereview.chromium.org/11818007

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@175701 4ff67af0-8c30-449e-8e8b-ad334ec8d88c