NaCl: distinguish between installed and hosted apps for load status UMA.
This will help determine if nexe loading problems are caused by networking
issues, or not.  A somewhat counter-intuitive aspect of this change is that
the classification of "installed app" vs. "not an installed app" may change at
different points in the load process, depending on where the manifest and nexe
are loaded from.  For example, if the manifest is loaded from a chrome-extension
URL, the load will be initially classified as an "installed app".  If the nexe
load then comes from an http URL, the classification will switch to "not an
installed app".  This means that the load status will be classified based on if
the last resource we attempted to load came from the disk or from the network.

BUG=171759

Review URL: https://chromiumcodereview.appspot.com/12041061

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@178677 4ff67af0-8c30-449e-8e8b-ad334ec8d88c