Report min (compile, download) as the progress event for PNaCl.
With dschuff's patch to LLC that limits the streaming buffer 
inside LLC so that the StreamBytes RPC blocks until
enough of the buffer is consumed.  Thus it never goes faster
than the compile.  StreamBytes also depends on bytes being
downloaded, so it never goes faster than the download.

This allows us to report compile progress if downloading is
faster than a compile, or report download progress if
compiling is faster.

BUG=none
TEST=browser_tests --gtest_filter=NaClBrowserTest*ProgressEvents* and manual

Review URL: https://chromiumcodereview.appspot.com/12093009

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@179605 4ff67af0-8c30-449e-8e8b-ad334ec8d88c