Only build one NaCl IRT for x86 (either 32-bit or 64-bit) on non Windows.
This saves some space on chromeos, where the ebuild file does
"doins ${FROM}/nacl_irt_*.nexe".  The linux installers also
do globbing, so could help with the linux install size.

For Windows, we need both because the chrome is still 32-bit
only, while NaCl can be 32-bit or 64-bit.  We can change
this once chrome is 64-bit on 64-bit systems.

I thought about changing the defaults in untrusted.gypi
to be more specific but did not. I assume that both flavors
are needed to build libraries for the SDK?


BUG=160988


Review URL: https://chromiumcodereview.appspot.com/12387020

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@185750 4ff67af0-8c30-449e-8e8b-ad334ec8d88c