Add PNaCl PPAPI browser tests via pre-translation.
Helps test that the shim and the compiler work together
correctly. At this moment, this does not use the in-browser
translator, since the ppapi browser test nexe is large and
the translation cache does not get shared between
invocations. See comments in bug 2484 about this.

BUG= https://code.google.com/p/nativeclient/issues/detail?id=2484
TEST= browser_tests --gtest_filter=*PNaCl*


Review URL: https://chromiumcodereview.appspot.com/12288013

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@186787 4ff67af0-8c30-449e-8e8b-ad334ec8d88c