Allow PNaCl NMF to set translator optimization options for experimentation.
This adds two fields:
(1) "-O": i
(2) "experimentation_flags": "arbitrary flags"

We may want to remove (2) soon... or just omit it.
The optimization flag is separated and made a uint so that
it can be tracked as a UMA stat more easily.

Also set default to -O0 for now.



TEST= browser_tests --gtest_filter=*Pnacl*Options*

This only tests that the plugin doesn't barf on those
new NMF options. To really test it we'd need to instrument
LLC to know that it's really getting these flags and get
output from LLC confirming that.  For now I have a manual
test:
http://www/~jvoung/IrrLicht/testirrnacl_portable.html
showing that it works though.

BUG= http://code.google.com/p/nativeclient/issues/detail?id=3325

Review URL: https://chromiumcodereview.appspot.com/12623004

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@187764 4ff67af0-8c30-449e-8e8b-ad334ec8d88c