Hookup RawVarData to SerializedVar
This hooks up the rewritten RawVarData implementation with SerializedVar. It also fixes tests to make sure that shmem ArrayBuffers are actually tested. Previously, no testing of shmem ArrayBuffers was happening because OS_LINUX is never defined in ppapi/tests and none of the other ArrayBuffer test sizes were above the threshhold to invoke that code path. This CL adds a PPB_Testing function to set the threshhold to a lower value, which permits this code path to be taken while still having fast tests.

BUG=

Review URL: https://chromiumcodereview.appspot.com/14208016

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@195189 4ff67af0-8c30-449e-8e8b-ad334ec8d88c