Pepper: Fix use-after-free bug in PluginVarTracker.
A PluginDispatcher may destroy itself in reaction to a channel error. However,
ProxyObjectVar instances may still have a pointer to the (destroyed)
dispatcher. This is problematic when the plugin wants to send a message to the
host to release the object; the ProxyObjectVar doesn't know the
PluginDispatcher has been deleted.

This change alters PluginDispatcher to invoke DidDeleteDispatcher() on the
PluginVarTracker, which will clean up the dangling dispatcher pointers.

BUG=233044

Review URL: https://chromiumcodereview.appspot.com/14054020

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@196850 4ff67af0-8c30-449e-8e8b-ad334ec8d88c