Remove inside_chromium_build gyp variable
In the days of olde, we supported compiling from inside a chromium checkout and
inside a WebKit checkout with some chromium libraries. In the latter configuration,
used mainly by the build.webkit.org bots, the paths to the WebKit headers relative
to chromium gyp files was different. This added a lot of complexity and some extra
scripts to fix up paths and includes. Blink only supports building within a chromium
checkout so all of this complexity can be removed.

Review URL: https://chromiumcodereview.appspot.com/14823005

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@199057 4ff67af0-8c30-449e-8e8b-ad334ec8d88c