Implement a V8 value<->PP_Var converter
NOTE: This patch is being relanded after it was reverted in 199944 due to a unittest breakage, fixed by 200279.

This implements a converter for transforming between V8 values and PP_Vars. This is needed to support transferring arrays/dictionaries (or arrays/objects in javascript) to and from the plugin using the pepper Post/HandleMessage APIs. The entire object graph is converted in the process.

TBR=darin@chromium.org

BUG=236958

Committed: https://src.chromium.org/viewvc/chrome?view=rev&revision=199938

Review URL: https://chromiumcodereview.appspot.com/14424006

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@200283 4ff67af0-8c30-449e-8e8b-ad334ec8d88c