Replace context parenting by sharing through mailboxes
Instead of tying contexts together through the parenting mechanism, we can allow
the client to produce the front buffer of an offscreen context into a mailbox,
and use that in the "parent" context.
It simplifies the code and the semantics wrt lost contexts, and we want to use
mailboxes anyway.

BUG=164095

Review URL: https://chromiumcodereview.appspot.com/15798014

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@204194 4ff67af0-8c30-449e-8e8b-ad334ec8d88c