Remove PP_HOSTRESOLVER_FLAGS_LOOPBACK_ONLY from PPB_HostResolver_Dev.
This was first introduced by PPB_HostResolver_Private. It mirrored net::HOST_RESOLVER_LOOPBACK_ONLY. However, it is actually a flag used by the host resolver implementation internally. It doesn't make sense to define this flag, because there is no way for a plugin to detect that only loopback addresses are configured and provide this information to the Pepper API.

BUG=247225
TEST=None

Review URL: https://chromiumcodereview.appspot.com/16831011

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@207354 4ff67af0-8c30-449e-8e8b-ad334ec8d88c