Pepper: Add FileRef messages for path information
Both FileIO and URLRequestInfo need to get path information for a FileRef resource (while in the renderer). This will introduce an extra round-trip to the browser in both paths.

It's possible that we'll do something more clever for FileIO (to go back to only one round trip), but I think we will need this for URLRequestInfo regardless.

BUG=225441

Review URL: https://chromiumcodereview.appspot.com/16975012

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@210394 4ff67af0-8c30-449e-8e8b-ad334ec8d88c