PNaCl: Simplify IRT wrapper a little by removing a global variable
There is no need for the temporary interface struct to be assembled in
a global variable, ppapi_hook.

This also removes a use of memcpy().  memcpy() is not available in
this context because this file isn't linked against libc.  It only
works because the compiler inlines the memcpy() call.

BUG=none
TEST=PNaCl tests in browser_tests

Review URL: https://codereview.chromium.org/20508002

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@213728 4ff67af0-8c30-449e-8e8b-ad334ec8d88c