NaCl CompletionCallback is called twice in WebSocket::ReceiveMessage and crashes the NaCl plugin
Some async APIs such as ReceiveMessage takes a completion callback which is called when the operation completes.
A completion callback must be called exactly once.

If a user calls an async API from non-main message loop, the completion callback will be post to the message loop from the main message loop.
But there was no way to detect whether PostRun of the completion callback was called and it caused repetitive calls of PostRun, which leads to an assertion failure and process crash.
We fix the issue in this CL.

BUG=258018
R=dmichael

Review URL: https://chromiumcodereview.appspot.com/19463005

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@213804 4ff67af0-8c30-449e-8e8b-ad334ec8d88c