Delete the unused CRX and webstore Zip modes from pnacl_component_crx_gen.
The PNaCl browser_tests currently only use the "installer"
mode, and that is the only one that works right now, because
that is the only one with the white-listed-chars-scrubbing.

For component updater deployment, I just use a shell script
that runs gyp 3X to compile the shims 3x, then zips it up.
See internal wiki for the script and HOWTO.

BUG=none, cleanup
TEST=remove out/Debug/pnacl && ...<build>... && browser_tests --gtest_filter=*Pnacl.SimpleLoad

Review URL: https://chromiumcodereview.appspot.com/21885006

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@215703 4ff67af0-8c30-449e-8e8b-ad334ec8d88c