Embed the PNACL_VERSION into the pnacl_public_pnacl_json file.
This allows us to read the version number for about:nacl for ChromeOS.

Normally, for component updater we have access to the
extension's manifest.json and we also encode the version in
the parent directory name. For ChromeOS we aren't using
component updater, so neither of those things is available
to us.  Put the version in a place that is readable.

Also move some files from gyp "sources" to "inputs" so that
touching the file will cause a rebuild.  The "sources"
list only affects what shows up in IDEs, and does not
affect the incremental build.

BUG=221381

Review URL: https://chromiumcodereview.appspot.com/22301014

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@216495 4ff67af0-8c30-449e-8e8b-ad334ec8d88c