Pepper: Fix FileSystemBrowserHost for isolated fs.
Currently the PepperFileSystemBrowserHost doesn't have a sane filesystem
context for isolated file systems. This was causing some crashes in an
application that used a CRX filesystem, which caused the FileRef change to be
reverted.

This was hard to debug because it was causing a browser crash; I cleaned up the
code in PepperInternalFileRefBackend to behave better when there's no file
system context.

TBR=jschuh
BUG=225441

Review URL: https://chromiumcodereview.appspot.com/22867028

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@219584 4ff67af0-8c30-449e-8e8b-ad334ec8d88c