Handle cache-control:no-store header in PNaCl translation cache
Pexe files with the cache-control:no-store header should not be cached.
Add a field to the PnaclCacheInfo struct, plumb the value all the way
from the plugin to the browser, and treat it basically the same way we
currently treat incognito translations (since we currently don't have an
off-the-record cache for those).
R=jvoung@chromium.org
BUG=none, noticed this was missing when doing cleanup

Review URL: https://chromiumcodereview.appspot.com/23458015

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@221275 4ff67af0-8c30-449e-8e8b-ad334ec8d88c