PPAPI: Make API changes w/o IDL changes Warnings if generators have changed
We frequently run in to the situation where we change the generators
which causes some real (but usually minor) change in header files, even
though the IDL is unchanged. We should just warn in that situation, so that
developers can use the CQ for these kind of generator changes.

BUG=

Review URL: https://chromiumcodereview.appspot.com/23620032

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@222190 4ff67af0-8c30-449e-8e8b-ad334ec8d88c