Return appropriate error codes for PPB_Flash_DRM
Currently we return PP_ERROR_FAILED for all error codes related to the flash
device ID. This returns codes that allow the caller to distinguish between
failures that occur because the user doesn't have access to the device ID
(e.g. because they have disabled it in their content setting or they are
in incognito mode) and other errors.

It also fixes an error code returned when the DRM interface isn't available
when calling GetVoucherFile.

BUG=

Review URL: https://codereview.chromium.org/25954004

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@228120 4ff67af0-8c30-449e-8e8b-ad334ec8d88c