[PPAPI] Fixed FileSystems from JavaScript not having a context.
Removed the "open existing file system" overload of
PepperFileSystemBrowserHost constructor, replacing it with a method
OpenExisting that asynchronously opens the file system, correctly
obtaining a FileSystemContext.

Rewrote PepperRendererConnection::OnMsgCreateResourceHostsFromHost to
deal with having to call an asynchronous function before sending the
reply to the plugin.

Re-enabled disabled test that was broken due to this bug and added
an additional regression test.

BUG=314007

Review URL: https://codereview.chromium.org/55133010

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@232605 4ff67af0-8c30-449e-8e8b-ad334ec8d88c