Split incompatible MSAA extensions.
This splits the commandbuffer client interfaces for extensions in the style of
GL_EXT_framebuffer_multisample (explicit resolve with glBlitFramebuffer)
vs. multisampled_render_to_texture.

glBlitFramebufferEXT() is also renamed to glBlitFrambufferCHROMIUM simply
to make it obvious that it is to be used with
glRenderbufferStorageMultisampleCHROMIUM rather than glRenderbufferStorageMultisampleEXT.

BUG=314214
R=brettw@chromium.org, kbr@chromium.org, piman@chromium.org
TBR=bsalomon@google.com

Review URL: https://codereview.chromium.org/63583002

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@234853 4ff67af0-8c30-449e-8e8b-ad334ec8d88c