Move some files from chrome/renderer/pepper to components/nacl/renderer/
Move pnacl_translation_resource_host files as they have no dependencies of the chrome layer.

Move the creation of the nacl private interface (ppb_nacl_private_impl) to the components directory : 
it requires to remove the IsOffTheRecord member of the interface (to not depend on chrome/)
but it's not used anyway.

This is part of an effort to componentize NaCl code.

BUG=244791

Review URL: https://codereview.chromium.org/88193002

git-svn-id: http://src.chromium.org/svn/trunk/src/ppapi@238083 4ff67af0-8c30-449e-8e8b-ad334ec8d88c